refactor: Improve ConfigDialog organization and visibility - Renamed PluginConfigDialog to ConfigDialog, added Debug derives, fixed imports in action modules
