+++ src/eg/action/group.rs
@@ -1,7 +1,7 @@
 use super::base::{ActionBase, ActionInfo};
 use crate::core::Error;
 use crate::core::event::Event;
-use crate::eg::classes::plugin_config::ConfigDialog;
+use crate::eg::classes::ConfigDialog;
 use uuid::Uuid;
 
 pub struct ActionGroup {
